feat: started on newsletter + fix: moved emailbox action to a different page oh boy surely this isn't going to cause any issues
