+++ src/routes/newsletter/+page.server.ts
@@ -1,0 +1,86 @@
+import {type Actions, fail} from "@sveltejs/kit";
+import {SECRET_IP_HASH_SALT} from "$env/static/private";
+import {createHash} from "node:crypto";
+import {PUBLIC_DEV} from "$env/static/public";
+
+export const prerender = false;
+
+export const actions = {
+    waitlist: async ({request, getClientAddress, locals: {supabase}}) => {
+        const formData = await request.formData();
+
+        // fake invisible field to prevent spam bots
+        const nickname = formData.get('nickname') as string;
+        if (nickname) return;
+
+        let email = formData.get('email') as string;
+        email = email.trim().toLowerCase();
+        const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+
+        if (!emailRegex.test(email)) {
+            return fail(400, {success: false, threat: 'Invalid email! Hmmmm...'});
+        }
+
+        // getting the user's ip for spam prevention
+        const clientIp = getClientAddress() || 'unknown';
+
+        // hashing it so even I can't read or use it even if I wanted to really hard
+        const SALT = SECRET_IP_HASH_SALT;
+
+        const hashIP = (ip: string) => {
+            return createHash('sha256').update(ip + SALT).digest('hex');
+        }
+
+        const hashedIP = hashIP(clientIp);
+
+        if (clientIp === 'unknown') {
+            // if unknown, bottleneck it to 300 just in case
+            // not that i'm gonna have that many users with a secure vpn
+            // or something but still
+            const {count: unCount, error: selError} = await supabase
+                .from('waitlist')
+                .select('*', {count: "exact", head: true})
+                .eq('hashed_ip', 'unknown');
+
+            const unCountZeroed = unCount || 0;
+
+            if (unCountZeroed > 500) {
+                return fail(400, {success: false, threat: 'Something went wrong? Are you per chance behind a proxy?'});
+            }
+        } else {
+            // checking if IP was used more than 6 times
+            const {count, error: selError} = await supabase
+                .from('waitlist')
+                .select('*', {count: "exact", head: true})
+                .eq('hashed_ip', hashedIP);
+
+            if(selError) {
+                if (PUBLIC_DEV) console.error(selError);
+                return fail(400, {success: false, threat: 'Oh no! Something went wrong! Try again!'});
+            }
+
+            const ipCountZeroed = count ?? 0;
+
+            if (ipCountZeroed > 6) {
+                return fail(400, {success: false, threat: 'Oh no! something went wrong! (using a VPN?)'});
+            }
+        }
+
+        // adding the user to the waitlist if everything is ok
+        const {error: inError} = await supabase
+            .from('waitlist')
+            .insert({email: email, hashed_ip: hashedIP})
+            .select()
+            .single();
+
+        if(inError) {
+            if (PUBLIC_DEV) console.error(inError);
+            if (inError.message === 'duplicate key value violates unique constraint "waitlist_email_key"') {
+                return fail(400, {success: false, threat: "You already signed up for it!"});
+            }
+            return fail(400, {success: false, threat: 'Oh no! Something went wrong! Try again!'});
+        }
+
+        return {success: true, threat: 'Signed up successfully! Thx! Stay tuned.',};
+    }
+} satisfies Actions;
